+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+
+-   Add a `--no-py` flag to the `east util cortex-debug`. It fixes possible
+    issues with RTT in VSCode.
+
 ## [0.18.1] - 2024-04-04
 
 ### Fixed